renamed package fakeFSO to fake_FSO
