+++ CommandLine/src/test/java/file_system_adapter/fake_FSO/FakeFileTest.java
@@ -1,4 +1,4 @@
-package file_system_adapter.fakeFSO;
+package file_system_adapter.fake_FSO;
 
 import org.junit.Before;
 import org.junit.Test;
